+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Add `VMStat` interface for Linux. [#59](https://github.com/elastic/go-sysinfo/pull/59)
+
 ### Changed
 
 ### Deprecated